@@ -4,16 +4,17 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at http://mozilla.org/MPL/2.0/.
 
-// +build go1.14,!noquic,!go1.17
+//go:build go1.15 && !noquic
+// +build go1.15,!noquic
 
 package connections
 
 import (
 	"crypto/tls"
 	"net"
+	"net/url"
 
 	"github.com/lucas-clemente/quic-go"
-	"github.com/syncthing/syncthing/lib/util"
 )
 
 var (
@@ -22,6 +23,17 @@ var (
 		KeepAlive:          true,
 	}
 )
+
+func quicNetwork(uri *url.URL) string {
+	switch uri.Scheme {
+	case "quic4":
+		return "udp4"
+	case "quic6":
+		return "udp6"
+	default:
+		return "udp"
+	}
+}
 
 type quicTlsConn struct {
 	quic.Session
@@ -47,24 +59,13 @@ func (q *quicTlsConn) Close() error {
 }
 
 func (q *quicTlsConn) ConnectionState() tls.ConnectionState {
-	qcs := q.Session.ConnectionState()
-	return tls.ConnectionState{
-		Version:                     qcs.Version,
-		HandshakeComplete:           qcs.HandshakeComplete,
-		DidResume:                   qcs.DidResume,
-		CipherSuite:                 qcs.CipherSuite,
-		NegotiatedProtocol:          qcs.NegotiatedProtocol,
-		NegotiatedProtocolIsMutual:  qcs.NegotiatedProtocolIsMutual,
-		ServerName:                  qcs.ServerName,
-		PeerCertificates:            qcs.PeerCertificates,
-		VerifiedChains:              qcs.VerifiedChains,
-		SignedCertificateTimestamps: qcs.SignedCertificateTimestamps,
-		OCSPResponse:                qcs.OCSPResponse,
-		TLSUnique:                   qcs.TLSUnique,
-	}
+	return q.Session.ConnectionState().TLS.ConnectionState
 }
 
-// Sort available packet connections by ip address, preferring unspecified local address.
-func packetConnLess(i interface{}, j interface{}) bool {
-	return util.AddressUnspecifiedLess(i.(net.PacketConn).LocalAddr(), j.(net.PacketConn).LocalAddr())
+func packetConnUnspecified(conn interface{}) bool {
+	// Since QUIC connections are wrapped, we can't do a simple typecheck
+	// on *net.UDPAddr here.
+	addr := conn.(net.PacketConn).LocalAddr()
+	host, _, err := net.SplitHostPort(addr.String())
+	return err == nil && net.ParseIP(host).IsUnspecified()
 }
