@@ -1,11 +1,10 @@
 // Copyright (C) 2014 The Protocol Authors.
 
+//go:build !windows && !darwin
 // +build !windows,!darwin
 
 package protocol
 
 // Normal Unixes uses NFC and slashes, which is the wire format.
 
-type nativeModel struct {
-	Model
-}
+func makeNative(m Model) Model { return m }
