@@ -4,11 +4,13 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
+//go:build ignore
 // +build ignore
 
 package main
 
 import (
+	"bufio"
 	"encoding/json"
 	"log"
 	"os"
@@ -20,12 +22,13 @@ import (
 )
 
 var trans = make(map[string]string)
-var attrRe = regexp.MustCompile(`\{\{'([^']+)'\s+\|\s+translate\}\}`)
-var attrReCond = regexp.MustCompile(`\{\{.+\s+\?\s+'([^']+)'\s+:\s+'([^']+)'\s+\|\s+translate\}\}`)
+var attrRe = regexp.MustCompile(`\{\{\s*'([^']+)'\s+\|\s+translate\s*\}\}`)
+var attrReCond = regexp.MustCompile(`\{\{.+\s+\?\s+'([^']+)'\s+:\s+'([^']+)'\s+\|\s+translate\s*\}\}`)
+var jsRe = regexp.MustCompile(`\$translate.instant\("([^"]+)"\)`)
 
 // exceptions to the untranslated text warning
 var noStringRe = regexp.MustCompile(
-	`^((\W*\{\{.*?\}\} ?.?\/?.?(bps)?\W*)+(\.stignore)?|[^a-zA-Z]+.?[^a-zA-Z]*|[kMGT]?B|Twitter|JS\W?|DEV|https?://\S+)$`)
+	`^((\W*\{\{.*?\}\} ?.?\/?.?(bps)?\W*)+(\.stignore)?|[^a-zA-Z]+.?[^a-zA-Z]*|[kMGT]?B|Twitter|JS\W?|DEV|https?://\S+|TechUi)$`)
 
 // exceptions to the untranslated text warning specific to aboutModalView.html
 var aboutRe = regexp.MustCompile(`^([^/]+/[^/]+|(The Go Pro|Font Awesome ).+|Build \{\{.+\}\}|Copyright .+ the Syncthing Authors\.)$`)
@@ -108,20 +111,46 @@ func walkerFor(basePath string) filepath.WalkFunc {
 			return err
 		}
 
-		if filepath.Ext(name) == ".html" && info.Mode().IsRegular() {
-			fd, err := os.Open(name)
-			if err != nil {
-				log.Fatal(err)
-			}
+		if !info.Mode().IsRegular() {
+			return nil
+		}
+		fd, err := os.Open(name)
+		if err != nil {
+			log.Fatal(err)
+		}
+		defer fd.Close()
+		switch filepath.Ext(name) {
+		case ".html":
 			doc, err := html.Parse(fd)
 			if err != nil {
 				log.Fatal(err)
 			}
-			fd.Close()
 			generalNode(doc, filepath.Base(name))
+		case ".js":
+			for s := bufio.NewScanner(fd); s.Scan(); {
+				for _, matches := range jsRe.FindAllStringSubmatch(s.Text(), -1) {
+					translation(matches[1])
+				}
+			}
 		}
 
 		return nil
+	}
+}
+
+func collectThemes(basePath string) {
+	files, err := os.ReadDir(basePath)
+	if err != nil {
+		log.Fatal(err)
+	}
+	for _, f := range files {
+		if f.IsDir() {
+			key := "theme-name-" + f.Name()
+			if _, ok := trans[key]; !ok {
+				name := strings.Title(f.Name())
+				trans[key] = name
+			}
+		}
 	}
 }
 
@@ -139,6 +168,7 @@ func main() {
 	var guiDir = os.Args[2]
 
 	filepath.Walk(guiDir, walkerFor(guiDir))
+	collectThemes(guiDir)
 
 	bs, err := json.MarshalIndent(trans, "", "   ")
 	if err != nil {
